requirements.txt
psycopg                                        
psycopg2-binary      
Flask
Gunicorn
